Remove ViewFinder as field and always use DefaultViewFinder
Move findIn from ViewFinder to DefaultViewFinder

Continue work of replacing findAllIn(many args) with findAllIn(context).
