+++ src/main/scala/com/atomist/rug/kind/rug/archive/RugArchiveProjectType.scala
@@ -14,11 +14,7 @@ class RugArchiveProjectType
   def description: String = "Rug archive"
 
   // Members declared in com.atomist.rug.kind.dynamic.ViewFinder
-  protected def findAllIn(rugAs: com.atomist.source.ArtifactSource,
-                          selected: com.atomist.rug.parser.Selected,
-                          context: TreeNode,
-                          poa: com.atomist.project.ProjectOperationArguments,
-                          identifierMap: Map[String,Object]): Option[Seq[com.atomist.rug.spi.MutableView[_]]] = {
+  override def findAllIn(context: TreeNode): Option[Seq[com.atomist.rug.spi.MutableView[_]]] = {
     context match {
       case pv: ProjectMutableView if pv.directoryExists(".atomist") =>
         Some(Seq(new RugArchiveProjectMutableView(pv)))
